refactor: pure TS file name's first character is supposed to be started by lowercase
