+++ src/components/Settings/ExtractorSettings.vue
@@ -29,8 +29,8 @@ import ColorSettings from './ColorSettings.vue'
 import { extractorMapper } from '@/store/modules/extractor'
 import Vue from 'vue'
 import { ExtractStrategy } from '@/domains/extractor'
-import SymbolExtractByArea from '@/domains/extractStrategies/SymbolExtractByArea'
-import LineExtract from '@/domains/extractStrategies/LineExtract'
+import SymbolExtractByArea from '@/domains/extractStrategies/symbolExtractByArea'
+import LineExtract from '@/domains/extractStrategies/lineExtract'
 import { axesMapper } from '@/store/modules/axes'
 import { canvasMapper } from '@/store/modules/canvas'
 import { datasetMapper } from '@/store/modules/dataset'
