+++ src/domains/extractStrategies/lineExtract.test.js
