+++ src/domains/extractStrategies/lineExtract.ts
