+++ src/domains/extractStrategies/symbolExtractByArea.test.js
