+++ src/domains/extractStrategies/symbolExtractByArea.ts
